arm sim class created (non-functional commit)
